+++ packages/bpk-scrim-utils/src/withScrim.js
@@ -30,11 +30,12 @@ import {
 } from './scroll-utils';
 
 import STYLES from './bpk-scrim-content.scss';
+import { onClosePropType } from './customPropTypes';
 
 const getClassName = cssModules(STYLES);
 
 const withScrim = WrappedComponent => {
-  class component extends Component {
+  class WithScrim extends Component {
     constructor() {
       super();
 
@@ -162,17 +163,18 @@ const withScrim = WrappedComponent => {
     }
   }
 
-  component.displayName = wrapDisplayName(WrappedComponent, 'withScrim');
+  WithScrim.displayName = wrapDisplayName(WrappedComponent, 'withScrim');
 
-  component.propTypes = {
-    onClose: PropTypes.func.isRequired,
+  WithScrim.propTypes = {
     getApplicationElement: PropTypes.func.isRequired,
+    onClose: onClosePropType,
     isIphone: PropTypes.bool,
     containerClassName: PropTypes.string,
     closeOnScrimClick: PropTypes.bool,
   };
 
-  component.defaultProps = {
+  WithScrim.defaultProps = {
+    onClose: null,
     isIphone: /iPhone/i.test(
       typeof window !== 'undefined' ? window.navigator.platform : '',
     ),
@@ -180,7 +182,7 @@ const withScrim = WrappedComponent => {
     closeOnScrimClick: true,
   };
 
-  return component;
+  return WithScrim;
 };
 
 export default withScrim;