+++ changelog.md
@@ -3,7 +3,10 @@
 ## Unreleased
 
 **Added:**
-- bpk-component-input
+- bpk-component-dialog:
+  - New web Dialog component, see https://backpack.github.io/components/web/dialogs/
+
+- bpk-component-input:
   - Added `clearable`, `onClear` and `clearButtonLabel` props. These allow inputs to have an optional clear button that appears when the input is focused.
 
 ## 2018-01-25 - New link styles